extracts flourish logic into own component
