+++ src/app/components/Embeds/OEmbed/index.tsx
@@ -1,13 +1,13 @@
 /* eslint-disable camelcase */
-import React, { useContext, useEffect, useRef } from 'react';
+import React, { useContext } from 'react';
 import pathOr from 'ramda/src/pathOr';
 import { ServiceContext } from '../../../contexts/ServiceContext';
 import { RequestContext } from '../../../contexts/RequestContext';
 import EmbedHtml from '../EmbedHtml';
 import EmbedError from '../EmbedError';
+import Flourish from './Flourish';
 
 export type OEmbedProps = {
-  type: string;
   oembed: {
     version: string;
     provider_name: string;
@@ -24,25 +24,7 @@ export type OEmbedProps = {
 const OEmbedLoader = ({ oembed }: OEmbedProps) => {
   const { translations } = useContext(ServiceContext);
   const { isAmp, canonicalLink } = useContext(RequestContext);
-  const { html, width, height, provider_name } = oembed;
-  const iframeRef = useRef<HTMLIFrameElement>(null);
-
-  useEffect(() => {
-    const onMessage = (event: MessageEvent) => {
-      const message = typeof event.data === 'string' && JSON.parse(event.data);
-      if (
-        iframeRef.current &&
-        message?.sender === 'Flourish' &&
-        message?.context === 'iframe.resize'
-      ) {
-        iframeRef.current.style.height = `${message.height}px`;
-      }
-    };
-
-    window.addEventListener('message', onMessage);
-
-    return () => window.removeEventListener('message', onMessage);
-  });
+  const { html, provider_name } = oembed;
 
   if (isAmp) {
     const errorMessage = pathOr(
@@ -69,19 +51,7 @@ const OEmbedLoader = ({ oembed }: OEmbedProps) => {
   }
 
   if (provider_name === 'Flourish') {
-    const iframeSrc = html.match(/(?<=src=").*?(?=[?"])/)?.[0];
-    return (
-      <iframe
-        src={`${iframeSrc}?auto=1`}
-        ref={iframeRef}
-        frameBorder="0"
-        scrolling="no"
-        height={height}
-        width={width}
-        style={{ width: '100%' }}
-        title="Interactive or visual content"
-      />
-    );
+    return <Flourish oembed={oembed} />;
   }
 
   return <EmbedHtml embeddableContent={html} />;
